requirements.txt
Flask==3.0.0
flask-cors==4.0.0
flask-sqlalchemy==3.1.1
flask-bcrypt==1.0.1
flask-caching==2.1.0
flask-mail==0.10.0
flask-limiter==3.5.0
PyJWT==2.8.0
requests==2.31.0
cryptography==41.0.7
python-dotenv==1.0.0
gunicorn==21.2.0

